interval bug fix. fromArray implementation
JBacon.interval was passing double the amount of time it should have
been on the first Event.Next. That has been corrected.

JBacon.fromArray works, but currently has an issue with possibly
sending the array elements out of order. This is due to the underlying
concurrent nature of JBacon, where all data pushes happen in threads. A
way to synchronize the array elements so that they get delivered in a
predictable manner needs to be found.
java.util.concurrent.SynchronousQueue possibly?