interval now uses TimeUnit conversion. Documentation
The interval passed to JBacon.interval is not in milliseconds anymore,
but any kind of time you wish to use that is defined in the TimeUnit
class. Just pass in the correct TimeUnit to correspond to your
interval, and everything else will be taken care of.

Documentation is now included for JBacon.interval.

The time delta pushed in the interval EventStream is always in
nanoseconds. Please convert as necessary.